UserAuthentication/requirements.txt
fastapi
uvicorn
pyjwt
passlib[bcrypt]
supabase
python-dotenv
pydantic[email]
python-dateutil
pytz